Rename reactions folder to autoruns
